00Sammeln/requirements.txt
mysql-connector-python
#logging
#datetime
#delorean
#feedparser
argparse
black